Refactor mahebot to osfbot and change files accordingly
